requirements.txt
fastapi>=0.104.1
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pillow>=10.1.0
torch>=2.1.0
torchvision>=0.16.0
transformers>=4.35.0
huggingface-hub>=0.17.3
numpy>=1.24.3
scipy>=1.11.4
opencv-python>=4.8.1.78
redis>=5.0.1
pyjwt>=2.8.0
python-dotenv>=1.0.0
pydantic>=2.5.0
